Title: a/b 출력, Time: 125ms, Memory: 24MB, Status: Wrong Answer - CodeTree

+++ 240212/a/b 출력/a-divide-b.py
@@ -1,4 +1,4 @@
-a,b=map(int, input().split())
+a,b=map(float, input().split())
 
 c=a/b
